Merged PR 17858: Allow End to End tests to be used against a clustered environment
This change allows a HOST variable to be defined on the 'Luminate TMV - Full End to End Tests' pipeline to enable the Full End to End Tests to be run against a clustered environment. This value should be the ip address of the master node in the cluster.

Note the preliminary required checks for this PR are failing for another reason with the pipeline which is being investigated and so I have published the PR anyway.

Related work items: #53205