Merged PR 24041: Fix today's FE2E failures
Added a fix to clear local storage before the tests that check the PATH toggle. Attempted a fix for the timetable test that failed this morning but it is just too flaky, so have demoted it to become a semi-manual test. We have good enough coverage from some of our newer timetable tests, so that this should be OK.

Related work items: #81826