Merged PR 24430: BDD scenario for 83105-trains-menu-consistency
- Added scenarios to test styling and consistency of trains list context menu.
- Added test for Unmatch/Rematch for scheduled train and Find Train and Match for unscheduled train.

**Issue 1**
The test has revealed that there are two different black colours for trainslist context menu items that have submenu options:
- no hover: #212529'
- hover: #16181b

The test has been coded to pass with these two different colour codes.

**Issue 2**
The test has revealed that the **Match** menu item is not disabled when the user does not have the role but is not present at all in the menu.

A separate bug has been raised for this issue as it is functional and not related to styling.

Related work items: #83105