Merged PR 25308: Added new method to InputBox.ts to support setting colour values vial the hex...
Added new method to InputBox.ts to support setting colour values via the hex-text input box within the colour picker rather than by the underlying form box which is encumbered by auto-complete JS triggers which introduce transient failures to the steps that use the method.

I was thinking, if we could put this in today and see if all is good on tomorrow morning's run, we could then roll out its use across the suite of tests as it executes a lot quicker that sending each character individually.

Related work items: #87476