Merged PR 26214: FE2E - remove TDD annotations - 51586 Current PunctualityUpdate - update punctuality time formats
This test was failing because some of the times that were precisely to the minute were expected to show seconds whereas they are formatted without seconds in the UI (e.g. 23:05:00 shows in the UI as 23:05, without the seconds part). Times with a non zero seconds component are shown with seconds  (e.g. 23:05:30 shows in the UI as 23:05:30, with the seconds part). Also a few of the punctuality times needed updating to get the test to pass.

Related work items: #93071